refactor: move internal/errors
