+++ services/wallet/responses/router_suggested_routes.go
@@ -1,7 +1,7 @@
 package responses
 
 import (
-	"github.com/status-im/status-go/errors"
+	"github.com/status-im/status-go/internal/errors"
 	"github.com/status-im/status-go/services/wallet/router/routes"
 )
