+++ services/wallet/responses/router_transactions.go
@@ -6,7 +6,7 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 
 	"github.com/status-im/status-go/crypto/types"
-	"github.com/status-im/status-go/errors"
+	"github.com/status-im/status-go/internal/errors"
 	"github.com/status-im/status-go/services/wallet/requests"
 	"github.com/status-im/status-go/services/wallet/wallettypes"
 )
