+++ services/wallet/routeexecution/errors.go
@@ -1,7 +1,7 @@
 package routeexecution
 
 import (
-	"github.com/status-im/status-go/errors"
+	"github.com/status-im/status-go/internal/errors"
 )
 
 // Abbreviation `W` for the error code stands for Wallet
