+++ services/wallet/router/router.go
@@ -12,7 +12,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 
-	"github.com/status-im/status-go/errors"
+	"github.com/status-im/status-go/internal/errors"
 	"github.com/status-im/status-go/internal/transactions"
 	"github.com/status-im/status-go/logutils"
 	communityToken "github.com/status-im/status-go/protocol/communities/token"
